+++ config.go
@@ -14,30 +14,7 @@ type Spec struct {
 	// Hostname is the container's host name.
 	Hostname string `json:"hostname"`
 	// Mounts profile configuration for adding mounts to the container's filesystem.
-	Mounts []Mount `json:"mounts"`
-	// Hooks are the commands run at various lifecycle events of the container.
-	Hooks Hooks `json:"hooks"`
-}
-
-type Hooks struct {
-	// Prestart is a list of hooks to be run before the container process is executed.
-	// On Linux, they are run after the container namespaces are created.
-	Prestart []Hook `json:"prestart"`
-	// Poststop is a list of hooks to be run after the container process exits.
-	Poststop []Hook `json:"poststop"`
-}
-
-// Mount specifies a mount for a container.
-type Mount struct {
-	// Type specifies the mount kind.
-	Type string `json:"type"`
-	// Source specifies the source path of the mount.  In the case of bind mounts on
-	// linux based systems this would be the file on the host.
-	Source string `json:"source"`
-	// Destination is the path where the mount will be placed relative to the container's root.
-	Destination string `json:"destination"`
-	// Options are fstab style mount options.
-	Options string `json:"options"`
+	MountPoints []MountPoint `json:"mounts"`
 }
 
 // Process contains information to start a specific application inside the container.
@@ -72,9 +49,10 @@ type Platform struct {
 	Arch string `json:"arch"`
 }
 
-// Hook specifies a command that is run at a particular event in the lifecycle of a container.
-type Hook struct {
-	Path string   `json:"path"`
-	Args []string `json:"args"`
-	Env  []string `json:"env"`
+// MountPoint describes a directory that may be fullfilled by a mount in the runtime.json.
+type MountPoint struct {
+	// Name is a unique descriptive identifier for this mount point.
+	Name string `json:"name"`
+	// Path specifies the path of the mount. The path and child directories MUST exist, a runtime MUST NOT create directories automatically to a mount point.
+	Path string `json:"path"`
 }
